+++ SolastaUnfinishedBusiness/Displays/CharacterDisplay.cs
@@ -12,17 +12,17 @@ internal static class CharacterDisplay
 
         using (UI.HorizontalScope())
         {
-            UI.ActionButton("Discord".Bold().Khaki(), BootContext.OpenDiscord, UI.Width(200));
-            UI.ActionButton("Wiki".Bold().Khaki(), BootContext.OpenWiki, UI.Width(200));
-            UI.ActionButton("Change History".Bold().Khaki(), BootContext.OpenChangeLog, UI.Width(200));
+            UI.ActionButton("Discord".Bold().Khaki(), BootContext.OpenDiscord, UI.Width((float)200));
+            UI.ActionButton("Wiki".Bold().Khaki(), BootContext.OpenWiki, UI.Width((float)200));
+            UI.ActionButton("Change History".Bold().Khaki(), BootContext.OpenChangeLog, UI.Width((float)200));
         }
 
         using (UI.HorizontalScope())
         {
             UI.ActionButton("<b>Donate:</b> GitHub Sponsors".Khaki(), BootContext.OpenDonateGithubSponsors,
-                UI.Width(200));
-            UI.ActionButton("<b>Donate:</b> Patreon".Khaki(), BootContext.OpenDonatePatreon, UI.Width(200));
-            UI.ActionButton("<b>Donate:</b> PayPal".Khaki(), BootContext.OpenDonatePayPal, UI.Width(200));
+                UI.Width((float)200));
+            UI.ActionButton("<b>Donate:</b> Patreon".Khaki(), BootContext.OpenDonatePatreon, UI.Width((float)200));
+            UI.ActionButton("<b>Donate:</b> PayPal".Khaki(), BootContext.OpenDonatePayPal, UI.Width((float)200));
         }
 
         UI.Label();
